Switch to TreeNode interface
Switch from a concrete TreeNode struct to an interface initialized by
the RPC system under test.

This enables each system to specify their preferred object, improving
benchmark accuracy by ensuring every system initializes and caches
objects that are ready for serialization, as opposed to having to copy
values from a common structure.

One example that is able to benefit from this is gRPC, which is able to
directly set the values of the request instead of having to go through
one indirection of setting values in a temporary structure and then on
the gRPC struct.

We're under the assumption that, in a production environment, the user
of the RPC system will already have the object in their preferred format
(because they already read it from a DB, received it from another conn,
built it from scratch, etc), and won't have to go through it again.